Edit operator.isPresent key
Instead of writing a method to check if an operator is present, the
presence is hard-coded to be false on initialization and when the
operate function is called. When an operator key is clicked, the
isPresent key's value is changed to true and no other operator key can
be clicked.